ChatBot Using Gradio/requirements.txt
# Requirements for Gemini ChatBot

# Flask - Web framework
gradio>=3.0.0

# Google Generative AI
google-generativeai>=0.5.0

# Python dotenv - Load environment variables
python-dotenv>=1.0.0
